+++ lib/util/pdf_object_helper.py
@@ -2,10 +2,14 @@
 Some methods to help with the direct manipulation/processing of PyPDF2's PdfObjects
 """
 from collections import namedtuple
-from PyPDF2.generic import IndirectObject, PdfObject
 
+from PyPDF2.generic import IndirectObject, PdfObject
+from rich.markup import escape
+
+from lib.util.adobe_strings import DANGEROUS_PDF_KEYS
 from lib.util.exceptions import PdfWalkError
 from lib.util.logging import log
+from lib.util.string_utils import get_label_style
 
 
 # In the case of easy key/value pairs the reference_key and the reference_address are the same but
@@ -14,6 +18,24 @@ from lib.util.logging import log
 #   e.g. a reference_key for a /Font labeled /F1 could be '/Resources' but the reference_address
 # might be '/Resources[/Font][/F1] if the /Font is a directly embedded reference instead of a remote one.
 PdfObjectRef = namedtuple('PdfObjectRef', ['reference_key', 'reference_address', 'pdf_obj'])
+
+# For printing SymlinkNodes
+SymlinkRepresentation = namedtuple('SymlinkRepresentation', ['text', 'style'])
+
+
+def get_symlink_representation(from_node, to_node) -> SymlinkRepresentation:
+    """Returns a tuple (symlink_text, style)"""
+    reference_key = str(to_node.get_reference_key_for_relationship(from_node))
+    pdf_instruction = reference_key.split('[')[0]  # In case we ended up with a [0] or similar
+
+    if pdf_instruction in DANGEROUS_PDF_KEYS:
+        symlink_style = 'red_alert'
+    else:
+        symlink_style = get_label_style(to_node.label) + ' dim'
+
+    symlink_str = f"{escape(reference_key)} [bright_white]=>[/bright_white]"
+    symlink_str += f" {escape(str(to_node.target))} [grey](Non Child Reference)[/grey]"
+    return SymlinkRepresentation(symlink_str, symlink_style)
 
 
 def pdf_object_id(pdf_object):
@@ -53,6 +75,6 @@ def get_references(obj: PdfObject, ref_key=None, ref_address=None) -> [PdfObject
     return return_list
 
 
-def does_list_have_any_references(_list):
+def does_list_have_any_references(_list) -> bool:
     """Return true if any element of _list is an IndirectObject"""
     return any(isinstance(item, IndirectObject) for item in _list)