+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Next Release
-* Add `--version` option
+
+# 1.2.0
+* Dramatic expansion in the `pdfalyzer`'s binary data scouring capabilities:
+   * Add `chardet` library guesses as to the encoding of all unknown byte sequences and ranks them from most to least likely
+   * Add attempted decodes of all backtick, frontslash, single, double, and guillemet quoted strings in font binaries
+   * Add decode attempts with `Windows-1252`, `UTF-7`, and `UTF-16` encodings
+   * Add `--suppress-decodes` to suppress attempted decodes of quoted strings in font binaries
+   * Cool art gets generated when you swarm a binaries quoted strings, which are mostly but not totally random
+* The `--font` option takes an optional argument to limit the output to a single font ID
+
+* Add `--limit-decodes` to suppress attempted decodes of quoted strings in font binaries over a certain length
 * Add `--surrounding` option to specify number of bytes to print/decode before and after suspicious bytes; decrease default number of surrounding bytes
-* Attempt Windows-1252 encoding for suspicious bytes
-* Fix scanning for several BOMs
+* Add `--version` option
+* `extract_guillemet_quoted_bytes()` and `extract_backtick_quoted_bytes()` are now iterators
+* Fix scanning for `UTF-16` BOM in font binary
 
 # 1.1.0
 * Print unprintable ascii characters in the font binary forced decode attempt with a bracket notation. e.g. print the string `[BACKSPACE]` instead of deleting the previous character